+++ src/test/java/org/scion/jpan/internal/AbstractSegmentsMinimalTest.java
@@ -17,33 +17,18 @@ package org.scion.jpan.internal;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
 
-import com.google.protobuf.ByteString;
-import com.google.protobuf.Timestamp;
 import java.nio.ByteBuffer;
-import java.time.Instant;
 import java.util.Arrays;
 import org.scion.jpan.ScionUtil;
 import org.scion.jpan.proto.control_plane.Seg;
-import org.scion.jpan.proto.crypto.Signed;
 import org.scion.jpan.proto.daemon.Daemon;
 import org.scion.jpan.testutil.MockControlServer;
+import org.scion.jpan.testutil.Scenario;
 
-/**
- * Test cases: <br>
- * A (-): srcISD == dstISD; src == dst; (same ISD, same AS)<br>
- * B (UP): srcISD == dstISD; dst == core; (same ISD, dst is core)<br>
- * C (DOWN): srcISD == dstISD; src == core; (same ISD, dst is core)<br>
- * D (CORE): srcISD == dstISD; src == core, dst == core; (same ISD, src/dst are cores)<br>
- * E (UP, DOWN): srcISD == dstISD; (same ISD, src/dst are non-cores)<br>
- * F (UP, CORE): srcISD != dstISD; dst == core; (different ISDs, dst is core)<br>
- * G (CORE, DOWN): srcISD != dstISD; src == core; (different ISDs, src is cores)<br>
- * H (UP, CORE, DOWN): srcISD != dstISD; (different ISDs, src/dst are non-cores)<br>
- * I (CORE): srcISD != dstISD; (different ISDs, src/dst are cores)
- */
 public abstract class AbstractSegmentsMinimalTest {
   protected static final String AS_HOST = "my-as-host.org";
   protected static final String CFG_MINIMAL = "topologies/minimal/";
-  protected static final long ZERO = ScionUtil.parseIA("0-0:0:0");
+  protected static final String CFG_DEFAULT = "topologies/scionproto-default/";
 
   /** ISD 1 - core AS */
   protected static final long AS_110 = ScionUtil.parseIA("1-ff00:0:110");
@@ -76,6 +61,15 @@ public abstract class AbstractSegmentsMinimalTest {
   protected static final long AS_211 = ScionUtil.parseIA("2-ff00:0:211");
 
   protected static MockControlServer controlServer;
+  protected final Scenario scenario;
+
+  protected AbstractSegmentsMinimalTest() {
+    this(CFG_MINIMAL);
+  }
+
+  protected AbstractSegmentsMinimalTest(String cfg) {
+    scenario = Scenario.readFrom(cfg);
+  }
 
   protected static void checkMetaHeader(
       ByteBuffer rawBB, int hopCount0, int hopCount1, int hopCount2) {
@@ -86,10 +80,10 @@ public abstract class AbstractSegmentsMinimalTest {
   protected static void checkInfo(ByteBuffer rawBB, int segmentId, int flags) {
     assertEquals(flags, rawBB.get()); // Info0 flags
     assertEquals(0, rawBB.get()); // Info0 etc
-    // TODO fix -> XOR SegID!
     if (flags != 0) {
       assertEquals(segmentId, ByteUtil.toUnsigned(rawBB.getShort())); // Info0 SegID
     } else {
+      // TODO fix -> XOR SegID! -> assert!
       rawBB.getShort();
     }
     assertNotEquals(0, rawBB.getInt()); // Info0 timestamp
@@ -152,56 +146,6 @@ public abstract class AbstractSegmentsMinimalTest {
     assertEquals(ofs, act.length);
   }
 
-  private static Seg.HopField buildHopField(int expiry, int ingress, int egress) {
-    ByteString mac = ByteString.copyFrom(new byte[] {1, 2, 3, 4, 5, 6});
-    return Seg.HopField.newBuilder()
-        .setExpTime(expiry)
-        .setIngress(ingress)
-        .setEgress(egress)
-        .setMac(mac)
-        .build();
-  }
-
-  private static Seg.HopEntry buildHopEntry(int mtu, Seg.HopField hf) {
-    return Seg.HopEntry.newBuilder().setIngressMtu(mtu).setHopField(hf).build();
-  }
-
-  private static Seg.ASEntry buildASEntry(long isdAs, long nextIA, int mtu, Seg.HopEntry he) {
-    Signed.Header header =
-        Signed.Header.newBuilder()
-            .setSignatureAlgorithm(Signed.SignatureAlgorithm.SIGNATURE_ALGORITHM_ECDSA_WITH_SHA256)
-            .setTimestamp(now())
-            .build();
-    Seg.ASEntrySignedBody body =
-        Seg.ASEntrySignedBody.newBuilder()
-            .setIsdAs(isdAs)
-            .setNextIsdAs(nextIA)
-            .setMtu(mtu)
-            .setHopEntry(he)
-            .build();
-    Signed.HeaderAndBodyInternal habi =
-        Signed.HeaderAndBodyInternal.newBuilder()
-            .setHeader(header.toByteString())
-            .setBody(body.toByteString())
-            .build();
-    Signed.SignedMessage sm =
-        Signed.SignedMessage.newBuilder().setHeaderAndBody(habi.toByteString()).build();
-    return Seg.ASEntry.newBuilder().setSigned(sm).build();
-  }
-
-  private static Seg.PathSegment buildPath(int id, Seg.ASEntry... entries) {
-    long now = Instant.now().getEpochSecond();
-    Seg.SegmentInformation info =
-        Seg.SegmentInformation.newBuilder().setSegmentId(id).setTimestamp(now).build();
-    Seg.PathSegment.Builder builder =
-        Seg.PathSegment.newBuilder().setSegmentInfo(info.toByteString());
-    //    for (Seg.ASEntry entry : entries) {
-    //      builder.addAsEntries(entry);
-    //    }
-    builder.addAllAsEntries(Arrays.asList(entries));
-    return builder.build();
-  }
-
   private static Seg.SegmentsResponse buildResponse(
       Seg.SegmentType type, Seg.PathSegment... paths) {
     Seg.SegmentsResponse.Builder replyBuilder = Seg.SegmentsResponse.newBuilder();
@@ -211,26 +155,47 @@ public abstract class AbstractSegmentsMinimalTest {
     return replyBuilder.build();
   }
 
-  private static Timestamp now() {
-    Instant now = Instant.now();
-    // TODO correct? Set nanos?
-    return Timestamp.newBuilder().setSeconds(now.getEpochSecond()).setNanos(now.getNano()).build();
-  }
-
   protected void addResponses() {
     addResponse110_1111();
     addResponse110_1112();
     addResponse110_1121();
     addResponse120_210();
 
-    // TODO redo the following ones?
-    // addResponse111_110();
     addResponse110_111();
     addResponse110_112();
     addResponse110_120();
     addResponse120_121();
     addResponse110_210();
     addResponse210_211();
+  }
+
+  protected void addResponsesScionprotoDefault() {
+    long AS_130 = ScionUtil.parseIA("1-ff00:0:130");
+
+    addUpDown(AS_111, AS_130);
+    addUpDown(AS_111, AS_120);
+
+    addUpDown(AS_112, AS_130);
+    addUpDown(AS_112, AS_120);
+
+    addCore(AS_130, AS_120);
+    addCore(AS_120, AS_130);
+  }
+
+  private void addCore(long local, long origin) {
+    for (Seg.PathSegment path : scenario.getSegments(local, origin)) {
+      controlServer.addResponse(
+          local, true, origin, true, buildResponse(Seg.SegmentType.SEGMENT_TYPE_CORE, path));
+    }
+  }
+
+  private void addUpDown(long leaf, long core) {
+    for (Seg.PathSegment path : scenario.getSegments(leaf, core)) {
+      controlServer.addResponse(
+          leaf, false, core, true, buildResponse(Seg.SegmentType.SEGMENT_TYPE_UP, path));
+      controlServer.addResponse(
+          core, true, leaf, false, buildResponse(Seg.SegmentType.SEGMENT_TYPE_DOWN, path));
+    }
   }
 
   private void addResponse110_111() {
@@ -249,52 +214,25 @@ public abstract class AbstractSegmentsMinimalTest {
     //        HopEntry: true mtu=1472
     //          HopField: exp=63 ingress=111 egress=0
 
-    Seg.HopEntry he00 = buildHopEntry(0, buildHopField(63, 0, 2));
-    Seg.ASEntry ase00 = buildASEntry(AS_110, AS_111, 1472, he00);
-    Seg.HopEntry he01 = buildHopEntry(1472, buildHopField(63, 111, 0));
-    Seg.ASEntry ase01 = buildASEntry(AS_111, ZERO, 1472, he01);
-    Seg.PathSegment path0 = buildPath(18215, ase00, ase01);
+    //    Seg.HopEntry he00 = buildHopEntry(0, buildHopField(63, 0, 2));
+    //    Seg.ASEntry ase00 = buildASEntry(AS_110, AS_111, 1472, he00);
+    //    Seg.HopEntry he01 = buildHopEntry(1472, buildHopField(63, 111, 0));
+    //    Seg.ASEntry ase01 = buildASEntry(AS_111, ZERO, 1472, he01);
+    //    Seg.PathSegment path0 = buildPath(18215, ase00, ase01);
 
+    Seg.PathSegment path0 = scenario.getSegments(AS_111, AS_110).get(0);
     controlServer.addResponse(
         AS_111, false, AS_110, true, buildResponse(Seg.SegmentType.SEGMENT_TYPE_UP, path0));
     controlServer.addResponse(
         AS_110, true, AS_111, false, buildResponse(Seg.SegmentType.SEGMENT_TYPE_DOWN, path0));
   }
 
-  //  private void addResponse110_111() {
-  //    //    Requesting segments: 1-ff00:0:110 -> 1-ff00:0:111
-  //    //    SEG: key=SEGMENT_TYPE_DOWN -> n=1
-  //    //    PathSeg: size=10
-  //    //    SegInfo:  ts=2024-01-05T12:53:41Z  id=50986
-  //    //      AS: signed=93   signature size=71
-  //    //      AS header: SIGNATURE_ALGORITHM_ECDSA_WITH_SHA256
-  // time=2024-01-05T12:53:41.919523006Z
-  //    //      AS Body: IA=1-ff00:0:110 nextIA=1-ff00:0:111  mtu=1472
-  //    //        HopEntry: true mtu=0
-  //    //          HopField: exp=63 ingress=0 egress=2
-  //    //      AS: signed=89   signature size=71
-  //    //      AS header: SIGNATURE_ALGORITHM_ECDSA_WITH_SHA256
-  // time=2024-01-05T12:53:43.935089190Z
-  //    //      AS Body: IA=1-ff00:0:111 nextIA=0-0:0:0  mtu=1472
-  //    //        HopEntry: true mtu=1472
-  //    //          HopField: exp=63 ingress=111 egress=0
-  //
-  //    Seg.HopEntry he00 = buildHopEntry(0, buildHopField(63, 0, 2));
-  //    Seg.ASEntry ase00 = buildASEntry(AS_110, AS_111, 0, he00);
-  //    Seg.HopEntry he01 = buildHopEntry(1472, buildHopField(63, 111, 0));
-  //    Seg.ASEntry ase01 = buildASEntry(AS_111, ZERO, 1472, he01);
-  //    Seg.PathSegment path0 = buildPath(50986, ase00, ase01);
-  //
-  //    controlServer.addResponse(
-  //        AS_110, true, AS_111, false, buildResponse(Seg.SegmentType.SEGMENT_TYPE_DOWN, path0));
-  //  }
-
   private void addResponse110_1111() {
     //  Requesting segments: 1-ff00:0:110 -> 1-ff00:0:1111
     //  SEG: key=SEGMENT_TYPE_DOWN -> n=1
     //  PathSeg: size=9
     //  SegInfo:  ts=2024-01-10T15:58:22Z  id=10619
-    //    AS Body: IA=1-ff00:0:110 nextIA=1-ff00:0:111  mtu=1472
+    //    AS Body: IA=1-ff00:0:110 nextIA=1-ff00:0:111  mtu=1460
     //      HopEntry: true mtu=0
     //        HopField: exp=63 ingress=0 egress=2
     //    AS Body: IA=1-ff00:0:111 nextIA=1-ff00:0:1111  mtu=1472
@@ -303,15 +241,7 @@ public abstract class AbstractSegmentsMinimalTest {
     //    AS Body: IA=1-ff00:0:1111 nextIA=0-0:0:0  mtu=1472
     //      HopEntry: true mtu=1472
     //        HopField: exp=63 ingress=123 egress=0
-
-    Seg.HopEntry he00 = buildHopEntry(0, buildHopField(63, 0, 2));
-    Seg.ASEntry ase00 = buildASEntry(AS_110, AS_111, 1472, he00);
-    Seg.HopEntry he01 = buildHopEntry(1472, buildHopField(63, 111, 1111));
-    Seg.ASEntry ase01 = buildASEntry(AS_111, AS_1111, 1472, he01);
-    Seg.HopEntry he02 = buildHopEntry(1472, buildHopField(63, 123, 0));
-    Seg.ASEntry ase02 = buildASEntry(AS_1111, ZERO, 1472, he02);
-    Seg.PathSegment path0 = buildPath(10619, ase00, ase01, ase02);
-
+    Seg.PathSegment path0 = scenario.getSegments(AS_1111, AS_110).get(0);
     controlServer.addResponse(
         AS_110, true, AS_1111, false, buildResponse(Seg.SegmentType.SEGMENT_TYPE_DOWN, path0));
     controlServer.addResponse(
@@ -323,7 +253,7 @@ public abstract class AbstractSegmentsMinimalTest {
     //  SEG: key=SEGMENT_TYPE_DOWN -> n=1
     //    PathSeg: size=10
     //      SegInfo:  ts=2024-01-10T17:11:53Z  id=25161
-    //      AS Body: IA=1-ff00:0:110 nextIA=1-ff00:0:111  mtu=1472
+    //      AS Body: IA=1-ff00:0:110 nextIA=1-ff00:0:111  mtu=1460
     //        HopEntry: true mtu=0
     //          HopField: exp=63 ingress=0 egress=2
     //      AS Body: IA=1-ff00:0:111 nextIA=1-ff00:0:1112  mtu=1472
@@ -332,15 +262,7 @@ public abstract class AbstractSegmentsMinimalTest {
     //      AS Body: IA=1-ff00:0:1112 nextIA=0-0:0:0  mtu=1472
     //        HopEntry: true mtu=1472
     //          HopField: exp=63 ingress=234 egress=0
-
-    Seg.HopEntry he00 = buildHopEntry(0, buildHopField(63, 0, 2));
-    Seg.ASEntry ase00 = buildASEntry(AS_110, AS_111, 1472, he00);
-    Seg.HopEntry he01 = buildHopEntry(1472, buildHopField(63, 111, 1112));
-    Seg.ASEntry ase01 = buildASEntry(AS_111, AS_1112, 1472, he01);
-    Seg.HopEntry he02 = buildHopEntry(1472, buildHopField(63, 234, 0));
-    Seg.ASEntry ase02 = buildASEntry(AS_1112, ZERO, 1472, he02);
-    Seg.PathSegment path0 = buildPath(25161, ase00, ase01, ase02);
-
+    Seg.PathSegment path0 = scenario.getSegments(AS_1112, AS_110).get(0);
     controlServer.addResponse(
         AS_110, true, AS_1112, false, buildResponse(Seg.SegmentType.SEGMENT_TYPE_DOWN, path0));
     controlServer.addResponse(
@@ -352,7 +274,7 @@ public abstract class AbstractSegmentsMinimalTest {
     //  SEG: key=SEGMENT_TYPE_DOWN -> n=1
     //    PathSeg: size=9
     //      SegInfo:  ts=2024-01-10T17:17:43Z  id=2700
-    //      AS Body: IA=1-ff00:0:110 nextIA=1-ff00:0:112  mtu=1472
+    //      AS Body: IA=1-ff00:0:110 nextIA=1-ff00:0:112  mtu=1460
     //        HopEntry: true mtu=0
     //          HopField: exp=63 ingress=0 egress=3
     //      AS Body: IA=1-ff00:0:112 nextIA=1-ff00:0:1121  mtu=1450
@@ -361,15 +283,7 @@ public abstract class AbstractSegmentsMinimalTest {
     //      AS Body: IA=1-ff00:0:1121 nextIA=0-0:0:0  mtu=1472
     //        HopEntry: true mtu=1472
     //          HopField: exp=63 ingress=345 egress=0
-
-    Seg.HopEntry he00 = buildHopEntry(0, buildHopField(63, 0, 3));
-    Seg.ASEntry ase00 = buildASEntry(AS_110, AS_112, 1472, he00);
-    Seg.HopEntry he01 = buildHopEntry(1472, buildHopField(63, 453, 1121));
-    Seg.ASEntry ase01 = buildASEntry(AS_112, AS_1121, 1450, he01);
-    Seg.HopEntry he02 = buildHopEntry(1472, buildHopField(63, 345, 0));
-    Seg.ASEntry ase02 = buildASEntry(AS_1121, ZERO, 1472, he02);
-    Seg.PathSegment path0 = buildPath(2700, ase00, ase01, ase02);
-
+    Seg.PathSegment path0 = scenario.getSegments(AS_1121, AS_110).get(0);
     controlServer.addResponse(
         AS_110, true, AS_1121, false, buildResponse(Seg.SegmentType.SEGMENT_TYPE_DOWN, path0));
     controlServer.addResponse(
@@ -384,7 +298,7 @@ public abstract class AbstractSegmentsMinimalTest {
     //    SegInfo:  ts=2024-01-05T15:02:17Z  id=5701
     //      AS: signed=93   signature size=71
     //      AS header: SIGNATURE_ALGORITHM_ECDSA_WITH_SHA256  time=2024-01-05T15:02:17.455400479Z
-    //      AS Body: IA=1-ff00:0:110 nextIA=1-ff00:0:112  mtu=1472
+    //      AS Body: IA=1-ff00:0:110 nextIA=1-ff00:0:112  mtu=1460
     //        HopEntry: true mtu=0
     //          HopField: exp=63 ingress=0 egress=3
     //      AS: signed=90   signature size=71
@@ -392,12 +306,7 @@ public abstract class AbstractSegmentsMinimalTest {
     //      AS Body: IA=1-ff00:0:112 nextIA=0-0:0:0  mtu=1450
     //        HopEntry: true mtu=1472
     //          HopField: exp=63 ingress=453 egress=0
-    Seg.HopEntry he00 = buildHopEntry(0, buildHopField(63, 0, 3));
-    Seg.ASEntry ase00 = buildASEntry(AS_110, AS_112, 1472, he00);
-    Seg.HopEntry he01 = buildHopEntry(1472, buildHopField(63, 453, 0));
-    Seg.ASEntry ase01 = buildASEntry(AS_112, ZERO, 1450, he01);
-    Seg.PathSegment path0 = buildPath(5701, ase00, ase01);
-
+    Seg.PathSegment path0 = scenario.getSegments(AS_112, AS_110).get(0);
     controlServer.addResponse(
         AS_110, true, AS_112, false, buildResponse(Seg.SegmentType.SEGMENT_TYPE_DOWN, path0));
   }
@@ -414,16 +323,10 @@ public abstract class AbstractSegmentsMinimalTest {
     //        HopField: exp=63 ingress=0 egress=10
     //    AS: signed=89   signature size=70
     //    AS header: SIGNATURE_ALGORITHM_ECDSA_WITH_SHA256  time=2024-01-10T12:48:17.672710479Z
-    //    AS Body: IA=1-ff00:0:110 nextIA=0-0:0:0  mtu=1472
+    //    AS Body: IA=1-ff00:0:110 nextIA=0-0:0:0  mtu=1460
     //      HopEntry: true mtu=1472
     //        HopField: exp=63 ingress=1 egress=0
-
-    Seg.HopEntry he00 = buildHopEntry(0, buildHopField(63, 0, 10));
-    Seg.ASEntry ase00 = buildASEntry(AS_120, AS_110, 1472, he00);
-    Seg.HopEntry he01 = buildHopEntry(1472, buildHopField(63, 1, 0));
-    Seg.ASEntry ase01 = buildASEntry(AS_110, ZERO, 1472, he01);
-    Seg.PathSegment path0 = buildPath(26755, ase00, ase01);
-
+    Seg.PathSegment path0 = scenario.getSegments(AS_110, AS_120).get(0);
     controlServer.addResponse(
         AS_110, true, AS_120, true, buildResponse(Seg.SegmentType.SEGMENT_TYPE_CORE, path0));
   }
@@ -443,13 +346,7 @@ public abstract class AbstractSegmentsMinimalTest {
     //    AS Body: IA=1-ff00:0:121 nextIA=0-0:0:0  mtu=1472
     //      HopEntry: true mtu=1472
     //        HopField: exp=63 ingress=104 egress=0
-
-    Seg.HopEntry he00 = buildHopEntry(0, buildHopField(63, 0, 21));
-    Seg.ASEntry ase00 = buildASEntry(AS_120, AS_121, 1472, he00);
-    Seg.HopEntry he01 = buildHopEntry(1472, buildHopField(63, 104, 0));
-    Seg.ASEntry ase01 = buildASEntry(AS_121, ZERO, 1472, he01);
-    Seg.PathSegment path0 = buildPath(48280, ase00, ase01);
-
+    Seg.PathSegment path0 = scenario.getSegments(AS_121, AS_120).get(0);
     controlServer.addResponse(
         AS_120, true, AS_121, false, buildResponse(Seg.SegmentType.SEGMENT_TYPE_DOWN, path0));
     controlServer.addResponse(
@@ -473,17 +370,12 @@ public abstract class AbstractSegmentsMinimalTest {
     //    AS Body: IA=1-ff00:0:120 nextIA=0-0:0:0  mtu=1472
     //      HopEntry: true mtu=1472
     //      HopField: exp=63 ingress=210 egress=0
-
-    Seg.HopEntry he00 = buildHopEntry(0, buildHopField(63, 0, 105));
-    Seg.ASEntry ase00 = buildASEntry(AS_210, AS_120, 1280, he00);
-    Seg.HopEntry he01 = buildHopEntry(1472, buildHopField(63, 210, 0));
-    Seg.ASEntry ase01 = buildASEntry(AS_120, ZERO, 1472, he01);
-    Seg.PathSegment path0 = buildPath(18204, ase00, ase01);
-
+    Seg.PathSegment path0 = scenario.getSegments(AS_120, AS_210).get(0);
+    Seg.PathSegment path1 = scenario.getSegments(AS_210, AS_120).get(0);
     controlServer.addResponse(
         AS_120, true, AS_210, true, buildResponse(Seg.SegmentType.SEGMENT_TYPE_CORE, path0));
     controlServer.addResponse(
-        AS_210, true, AS_120, true, buildResponse(Seg.SegmentType.SEGMENT_TYPE_CORE, path0));
+        AS_210, true, AS_120, true, buildResponse(Seg.SegmentType.SEGMENT_TYPE_CORE, path1));
   }
 
   private void addResponse110_210() {
@@ -503,18 +395,10 @@ public abstract class AbstractSegmentsMinimalTest {
     //        HopField: exp=63 ingress=210 egress=10
     //    AS: signed=88   signature size=70
     //    AS header: SIGNATURE_ALGORITHM_ECDSA_WITH_SHA256  time=2024-01-10T12:59:51.171109823Z
-    //    AS Body: IA=1-ff00:0:110 nextIA=0-0:0:0  mtu=1472
+    //    AS Body: IA=1-ff00:0:110 nextIA=0-0:0:0  mtu=1460
     //      HopEntry: true mtu=1472
     //        HopField: exp=63 ingress=1 egress=0
-
-    Seg.HopEntry he00 = buildHopEntry(0, buildHopField(63, 0, 105));
-    Seg.ASEntry ase00 = buildASEntry(AS_210, AS_120, 1280, he00);
-    Seg.HopEntry he01 = buildHopEntry(1472, buildHopField(63, 210, 10));
-    Seg.ASEntry ase01 = buildASEntry(AS_120, AS_110, 1472, he01);
-    Seg.HopEntry he02 = buildHopEntry(1472, buildHopField(63, 1, 0));
-    Seg.ASEntry ase02 = buildASEntry(AS_110, ZERO, 1472, he02);
-    Seg.PathSegment path0 = buildPath(15767, ase00, ase01, ase02);
-
+    Seg.PathSegment path0 = scenario.getSegments(AS_110, AS_210).get(0);
     controlServer.addResponse(
         AS_110, true, AS_210, true, buildResponse(Seg.SegmentType.SEGMENT_TYPE_CORE, path0));
   }
@@ -534,13 +418,7 @@ public abstract class AbstractSegmentsMinimalTest {
     //    AS Body: IA=2-ff00:0:211 nextIA=0-0:0:0  mtu=1472
     //      HopEntry: true mtu=1472
     //        HopField: exp=63 ingress=503 egress=0
-
-    Seg.HopEntry he00 = buildHopEntry(0, buildHopField(63, 0, 450));
-    Seg.ASEntry ase00 = buildASEntry(AS_210, AS_211, 1280, he00);
-    Seg.HopEntry he01 = buildHopEntry(1472, buildHopField(63, 503, 0));
-    Seg.ASEntry ase01 = buildASEntry(AS_211, ZERO, 1472, he01);
-    Seg.PathSegment path0 = buildPath(15299, ase00, ase01);
-
+    Seg.PathSegment path0 = scenario.getSegments(AS_211, AS_210).get(0);
     controlServer.addResponse(
         AS_210, true, AS_211, false, buildResponse(Seg.SegmentType.SEGMENT_TYPE_DOWN, path0));
     controlServer.addResponse(